refactor: update imports to use new visualizer types and remove shared visualizerTypes
