+++ src/test/config/defaults.test.ts
@@ -1,6 +1,6 @@
 import { DialectType } from '../../constants';
 import { DEFAULT_GCODE_CONFIG } from '../../config';
-import { ProjectionMode } from '../../shared/visualizerTypes';
+import { ProjectionMode } from '../../visualizer/types';
 
 describe('DEFAULT_GCODE_CONFIG', () => {
   // ---------------------------------------------------------------------------
