+++ src/test/providers/FanucProvider.test.ts
@@ -1,7 +1,7 @@
 import { TextDocument } from 'vscode-languageserver-textdocument';
 
 import { DialectType } from '../../constants';
-import { FormatterConfig } from '../../config/types';
+import { FormatterConfig } from '../../formatter/types';
 import { DocumentFormattingProvider } from '../../providers/DocumentFormattingProvider';
 import { FormatterService } from '../../providers/FormatterService';
 import { Range } from '../../parser/nodes';
