+++ src/visualizer/GCodeInterpreter.ts
@@ -26,10 +26,10 @@ import {
   WhileStatementNode,
 } from '../parser/nodes';
 import { ProgramNode } from '../parser/nodes/ProgramNode';
-import { DEFAULT_GCODE_CONFIG, InterpreterConfig } from '../config';
+import { DEFAULT_GCODE_CONFIG } from '../config';
 import { normalizeCommand } from '../utils/GCodeNormalizer';
 import { GCodeExpressionEvaluator } from './GCodeExpressionEvaluator';
-import { MotionHandler } from './types';
+import { InterpreterConfig, MotionHandler } from './types';
 
 /**
  * G-code Group 1 modal motion commands.
