+++ tsconfig.webview.json
@@ -14,6 +14,6 @@
     "outDir": "./dist/webview",
     "rootDir": "./src"
   },
-  "include": ["src/webview/**/*.ts", "src/shared/**/*.ts"],
+  "include": ["src/webview/**/*.ts"],
   "exclude": ["node_modules"]
 }
